A couple updates
- Fixed a couple mismatched values between the manual and the module
- Added a TP autosolver
- Added Gamepad/controller support
- Button press sounds now come directly from the buttons themselves
- Manual is now bundled with the module
