+++ Manual/js/ruleseed.js
