+++ Manual/pdfs/NeedyModules/.gitignore
@@ -1,0 +1,1 @@
+#Just to add empty folders
